# routes/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Music Stram descritption",
        "version": "1.0.0",
        "title": "Music Stream",
        "termsOfService": "",
        "contact": {
            "email": "hspharwinder@gmail.com"
        }
    },
    "host": "",
    "basePath": "/",
...
    "schemes": [
        "http",
        "https"
    ],
    "paths": {  
...
        "/allApprovedArtist": {
            "get": {
                "tags": [
                    "Approval"
                ],
                "summary": "Get all approved artist",
                "description": "Return all aprpoved artist",
                "operationId": "allApprovedArtist",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "responses": {
                    "200": {
                        "description": "Success, responsed json Array",
                        "schema": {
                            "$ref": "#/definitions/Artist"
                        }
                    },
                    "400": {
                        "description": "Fail to retreive record"
                    }
                }
            }
        },
        "/allPendingArtist": {
            "get": {
                "tags": [
                    "Approval"
                ],
                "summary": "Get all pending artist from approval",
                "description": "Return all pending artist from approval",
                "operationId": "allPendingArtist",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "responses": {
                    "200": {
                        "description": "Success, responsed json Array",
                        "schema": {
                            "$ref": "#/definitions/Artist"
                        }
                    },
                    "400": {
                        "description": "Fail to retreive record"
                    }
                }
            }
        },
        "/approveToArtist": {
            "post": {
                "tags": [
                    "Approval"
                ],
                "summary": "Give approval to artist, by admin",
                "description": "Give approval to artist by updating username and passord to the artist, process will be done by admin only",
                "operationId": "approveToArtist",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "formData",
                        "name": "artistId",
                        "description": "Artist id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "in": "formData",
                        "name": "userName",
                        "description": "User name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "password",
                        "description": "Password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "responsed json Array of user",
                        "schema": {
                            "$ref": "#/definitions/Artist"
                        }
                    },
                    "400": {
                        "description": "Fail to update record"
                    }
                }
            }
        },
        "/changeStatus": {
            "post": {
                "tags": [
                    "Approval"
                ],
                "summary": "Change status of any record active/in-active",
                "description": "Change status of any record based on record Id",
                "operationId": "changeStatus",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "formData",
                        "name": "id",
                        "description": "",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "in": "formData",
                        "name": "userName",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "status",
                        "description": "",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "responsed json Array of user",
                        "schema": {
                            "$ref": "#/definitions/Artist"
                        }
                    },
                    "400": {
                        "description": "Fail to update record"
                    }
                }
            }
        },
...
        "/fetchLikeDislike": {
            "post": {
                "tags": [
                    "Liking"
                ],
                "summary": "Fetch like dislike based on userId and media id",
                "description": "",
                "operationId": "fetchlikeDislike",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "formData",
                        "name": "userId",
                        "description": "",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "in": "formData",
                        "name": "mediaId",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "0/1  (0-dislike, 1-like)"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "petstore_auth": {
            "type": "oauth2",
            "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
            "flow": "implicit",
            "scopes": {
                "write:pets": "modify pets in your account",
                "read:pets": "read your pets"
            }
        },
        "api_key": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        }
    },
    "definitions": {  
...
        "Artist": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "type": {
                    "type": "integer",
                    "format": "int(11)"
                },
                "image": {
                    "type": "string",
                    "description": "First upload your image through 'filePost' endpoints and paste here responed image path"
                },
                "status": {
                    "type": "integer",
                    "format": "tinyInt"
                },
                "phone_no": {
                    "type": "string",
                    "format": "mobile"
                },
                "description": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "Artist"
            }
        },
...
    },
...
}